Chore: 최상위 파일 변경 최상위 파일 domain, global로 변경 Resolves:#3
