+++ src/main/java/umc/duckmelang/domain/temp/controller/TempRestController.java
@@ -1,4 +1,4 @@
-package umc.duckmelang.temp.controller;
+package umc.duckmelang.domain.temp.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
@@ -6,10 +6,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import umc.duckmelang.apipayload.ApiResponse;
-import umc.duckmelang.temp.converter.TempConverter;
-import umc.duckmelang.temp.dto.TempResponse;
-import umc.duckmelang.temp.service.TempQueryService;
+import umc.duckmelang.global.apipayload.ApiResponse;
+import umc.duckmelang.domain.temp.converter.TempConverter;
+import umc.duckmelang.domain.temp.dto.TempResponse;
+import umc.duckmelang.domain.temp.service.TempQueryService;
 
 @RestController
 @RequestMapping("/temp")
