+++ src/main/java/umc/duckmelang/domain/temp/dto/TempRequest.java
@@ -1,0 +1,4 @@
+package umc.duckmelang.domain.temp.dto;
+
+public class TempRequest {
+}
