+++ src/main/java/umc/duckmelang/domain/temp/service/TempCommandService.java
@@ -1,0 +1,4 @@
+package umc.duckmelang.domain.temp.service;
+
+public interface TempCommandService {
+}
