+++ src/main/java/umc/duckmelang/global/apipayload/code/BaseCode.java
@@ -1,4 +1,4 @@
-package umc.duckmelang.apipayload.code;
+package umc.duckmelang.global.apipayload.code;
 
 public interface BaseCode {
     ReasonDTO getReason();
