+++ src/main/java/umc/duckmelang/global/apipayload/code/ErrorReasonDTO.java
@@ -1,4 +1,4 @@
-package umc.duckmelang.apipayload.code;
+package umc.duckmelang.global.apipayload.code;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
-public class ReasonDTO {
+public class ErrorReasonDTO {
     private HttpStatus httpStatus;
 
     private final boolean isSuccess;
