+++ src/main/java/umc/duckmelang/global/apipayload/code/ReasonDTO.java
@@ -1,4 +1,4 @@
-package umc.duckmelang.apipayload.code;
+package umc.duckmelang.global.apipayload.code;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @Builder
-public class ErrorReasonDTO {
+public class ReasonDTO {
     private HttpStatus httpStatus;
 
     private final boolean isSuccess;
