+++ src/main/java/umc/duckmelang/global/apipayload/code/status/ErrorStatus.java
@@ -1,10 +1,10 @@
-package umc.duckmelang.apipayload.code.status;
+package umc.duckmelang.global.apipayload.code.status;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
-import umc.duckmelang.apipayload.code.BaseErrorCode;
-import umc.duckmelang.apipayload.code.ErrorReasonDTO;
+import umc.duckmelang.global.apipayload.code.BaseErrorCode;
+import umc.duckmelang.global.apipayload.code.ErrorReasonDTO;
 
 @Getter
 @AllArgsConstructor
