+++ src/main/java/umc/duckmelang/global/apipayload/exception/handler/TempHandler.java
@@ -1,0 +1,11 @@
+package umc.duckmelang.global.apipayload.exception.handler;
+
+import umc.duckmelang.global.apipayload.code.BaseErrorCode;
+import umc.duckmelang.global.apipayload.exception.GeneralException;
+
+public class TempHandler extends GeneralException {
+
+    public TempHandler(BaseErrorCode errorCode) {
+        super(errorCode);
+    }
+}
