--- src/main/java/umc/duckmelang/temp/dto/TempRequest.java
@@ -1,4 +1,0 @@
-package umc.duckmelang.temp.dto;
-
-public class TempRequest {
-}
