--- src/main/java/umc/duckmelang/temp/service/TempCommandService.java
@@ -1,4 +1,0 @@
-package umc.duckmelang.temp.service;
-
-public interface TempCommandService {
-}
